feat: 강의 수정 기능 구현

+++ src/actions/course.ts
@@ -5,13 +5,11 @@ import { getSession } from "@/lib/auth";
 import { CourseFormSchema } from "@/schema/course";
 
 export async function addCourse(formData: any) {
-  // Return early if not admin
   const userSession = await getSession();
-  // TODO: Decrypt token here and check expiry & admin role
-  if (!userSession) return { isValid: false };
+  if (!userSession.isLoggedIn || !userSession.token) return { isValid: false };
 
   const validatedForm = CourseFormSchema.safeParse(formData);
-  console.log(validatedForm);
+
   if (!validatedForm.success) {
     return { errors: validatedForm.error.flatten().fieldErrors };
   }
@@ -21,7 +19,33 @@ export async function addCourse(formData: any) {
     try {
       const res = await api.post("/api/v1/course", data, userSession.token);
       const body = await res.json();
-      console.log(body);
+      if (body.code === 200) {
+        return { isValid: true, data: body.data };
+      }
+      return { isValid: false, errors: { unknown: [`* ${body.message}`] } };
+    } catch (e) {
+      console.log(e);
+    }
+  }
+  return { isValid: false };
+}
+
+export async function updateCourse(courseId: string, formData: any) {
+  const userSession = await getSession();
+  if (!userSession.isLoggedIn || !userSession.token) return { isValid: false };
+
+  const validatedForm = CourseFormSchema.safeParse(formData);
+
+  if (!validatedForm.success) {
+    return { errors: validatedForm.error.flatten().fieldErrors };
+  }
+
+  if (validatedForm.success) {
+    const data = validatedForm.data;
+    try {
+      console.log("??????????????????????????????????????????????????????????????????")
+      const res = await api.put(`/api/v1/course/${courseId}`, data, userSession.token);
+      const body = await res.json();
       if (body.code === 200) {
         return { isValid: true, data: body.data };
       }
@@ -34,10 +58,8 @@ export async function addCourse(formData: any) {
 }
 
 export async function deleteCourse(courseId: string) {
-  // Return early if not admin
   const userSession = await getSession();
-  // TODO: Decrypt token here and check expiry & admin role
-  if (!userSession) return { isValid: false };
+  if (!userSession.isLoggedIn || !userSession.token) return { isValid: false, status: "loggedOut" };
 
   try {
     const res = await api.delete(`/api/v1/course/${courseId}`, userSession.token);
